refactor: onAudioUpload関数とscheduledTask関数をtriggersディレクトリに移動
