+++ apps/functions/src/triggers/scheduledTask.ts
